refactor(sections): extract the line and column primitives into layout.rs
Pure motion: tab-stop indentation, cells/fields, and the multi-column
recurrence thresholds, with the tests that pin them.
